refactor: Move ODataQuery methods into ODataQueryBase
